Add a more accurate cputime count based on xhprofs implementation
We use the xhprof implementation cycle_time to get the cputime. This
reads the RTDSC register and calculates the elapsed time based on the
current CPU frequency. This requires the process to be bound to a CPU,
otherwise the OS might preempt our process and migrate it to another CPU.
As CPU's do not start up at the very same moment, the clock counts usually
differ from cpu to cpu, leading to incorrect results.

Please note that this implementation might return wrong values on CPU
with speed stepping.

Signed-off-by: David Soria Parra <david.soria_parra@mayflower.de>